Don't both change directory and set GIT_WORK_TREE
We used to both change the working directory and set GIT_WORK_TREE
when calling git merge-recursive. This works with older versions of
git (up to and including 1.5.4.x, or thereabouts) because
merge-recursive ignores GIT_WORK_TREE. It stopped doing that sometime
just before 1.5.5, which broke StGit. (Given both GIT_WORK_TREE=".."
and cwd "..", it assumed the work tree was at "../..", instead of just
".." like we intended. This made t2800 fail.)

Just dropping the cd would solve the problem for new gits, but would
break StGit with older gits. But it works in both cases if we keep the
cd, and unconditionally set GIT_WORK_TREE=".". So this patch does
precisely that.

(git read-tree -u -m is also called from IndexAndWorktree, and it
 seems to always ignore GIT_WORK_TREE just like merge-recursive used
 to do. The patch fixes all calls made from IndexAndWorktree, so we're
 good if read-tree starts heeding GIT_WORK_TREE in the future.)

Signed-off-by: Karl Hasselström <kha@treskal.com>