Try "stg sink" without applied patches
It doesn't work, neither with an implicit nor an explicit patch to
sink. This is bug 11887 in the bug tracker.

(The implicit sink testcase actually passes, but that's just because
the test suite can't distinguish between a program bug and an orderly
abort.)

The test was adapted from the script attached to the bug report,
written by Erik Sandberg.

Signed-off-by: Karl Hasselström <kha@treskal.com>