Test operations on hidden patches
Previously, we didn't test this at all. Now we have some tests at
least; and, not surprisingly, a few of them fail.

The expected result of the tests are based on the following idea of
how hidden patches should work: They should behave just like unapplied
patches when named explicitly; but if not named explicitly, they
should behave as if they didn't exist. So for example, a push without
arguments should never push a hidden patch, but it should be possible
to push a hidden patch if it is named explicitly.

Signed-off-by: Karl Hasselström <kha@treskal.com>