Don't use the dashed for of git commands
With the release of git 1.6.0, the dashed forms of the git commands
are no longer installed in $PATH by default, so trying to use them is
likely to fail. This patch replaces all remaining dashed git calls by
their dashless equivalents -- even in comments and docs, for
consistency and to aid grepping.

Signed-off-by: Karl Hasselström <kha@treskal.com>