Auto-generate man pages for all StGit commands
Auto-generate man pages based on the docs that are in each
stgit/commands/<cmd>.py file. That doc format is extended in order to
support both brief command help output and manpage text.

The existing interactive help is made briefer (just the usage, the
one-line explanation, and the option listing).

The text in the man pages needs some love, both content-wise and
asciidoc-markup-wise. This is left for future patches to resolve. (It
should be rather simple to improve the man pages one by one now that
the infrastructure is in place.)

Signed-off-by: Karl Hasselström <kha@treskal.com>