Refactor stgit.commands.edit
Reorganize a few existing functions, and break out stuff from the main
function into subroutines.

While we're at it, move one of the old and all of the new functions to
stgit.lib.edit, so that we can use them in a later patch to implement
"stg refresh --edit".

This fixes one of the known failures in t3300-edit.

Signed-off-by: Karl Hasselström <kha@treskal.com>