Add release notes file
The idea is that the release announcement for the upcoming version of
StGit is prepared here. That way, the work of putting it together and
polishing it can be shared.

The text added by this patch was taken verbatim from the v0.15-rc1
announcement that Catalin posted to the git mailing list on July 12,
2009. The patch adding it to StGit (including this commit message) is
by Karl Wiberg <kha@treskal.com>.