squash: Make commit message editing more convenient
Very often, the commit message you want when squashing is the message
of the first patch. So instead of simply concatenating the messages of
all the patches, put a comment delimiter after the first one, and
ignore everything after the delimiter when reading the message back
in. If the user wants to use any part of the commented-out messages,
she can move text around however she wants, including deleting the
comment delimiter.

Signed-off-by: Karl Wiberg <kha@treskal.com>