When reading a config value, pick the last value, not the first
When reading the config values, we save all values for a given key in
a list, in the order we see them. Then, when asked for one value, we
used to return the _first_ value in the list. But the correct thing to
do in order to allow local configs (like the repository's .git/config)
to override global configs (like ~/.gitconfig) is to return the _last_
value.

Signed-off-by: Karl Wiberg <kha@treskal.com>