templates: fix attachinline option to generate proper mbox
There was a bug where the attachinline option to stg mail was
not spitting out a patch that can be used by git am, because the patch
file only contained the diff.

This modifies the template to include pretty much the same text in the
mail as is in the attachment.

Signed-off-by: Jesse Brandeburg <jesse.brandeburg@intel.com>
Signed-off-by: Peter P Waskiewicz Jr <peter.p.waskiewicz.jr@intel.com>