mail: don't hardcode space locations in template variables
This patch modifies stgit-mail in order to enable templates to move the ordering
of PATCH, the prefix, the version, and the number. The issue is due to stgit
hardcoding a ' ' after the previx, before version and before numbers. This means
that you can't design a template which re-orders these as it will result in
incorrect spacing being generated. To this end, introduce %(vspace), %(nspace)
and %(pspace) which allows a template to move the spacing around and enable
correct alignment inside a template. This won't change the default behavior (if
sticking to the standard templates) but enables more ability to customize and
modify the default behaviors.

Signed-off-by: Jacob Keller <jacob.e.keller@intel.com>
Signed-off-by: Peter P Waskiewicz Jr <peter.p.waskiewicz.jr@intel.com>