Run the test suite in parallel
On a 4-core test machine with the test scratch dir on NFS, this
reduces the time for running the complete suite from 7-8 minutes to 53
seconds, very close to a 10x speedup. On a similar machine with the
test scratch dir on an SSD, it reduces the time from 3:43 to 49
seconds, a respectable 4.5x speedup.

Signed-off-by: Karl Wiberg <kha@treskal.com>