test-lib.sh: Don't set GIT_CONFIG to a relative path
If we do, things will fail if we call any git commands from a
subdirectory of a repository, where the relative path ".git/config"
isn't the path to the config file. The following tests are fixed by
this patch:

   t1205-push-subdir.sh
   t1800-import.sh
   t2300-refresh-subdir.sh
   t2800-goto-subdir.sh

I'm not sure why this used to work. I'm guessing git has gotten more
pedantic (which is a fine decision---if we give it the path to the
config file and that path doesn't work, we want it to fail!), but I
haven't attempted to verify this.

Signed-off-by: Karl Wiberg <kha@treskal.com>