Honour PYTHON setting in all Makefiles
The top level Makefile allows the user to specify their Python using
the PYTHON variable; however the sub-Makefiles weren't honouring this,
which meant that if you didn't have a 'python2' executable then the
make 'all' and 'install' targets would work but 'doc' and 'install-doc'
would not.

Add explicit uses of $(PYTHON) everywhere we run a python script from
the Makefiles, bringing the sub-makefiles in line with the top level one.

Signed-off-by: Peter Maydell <peter.maydell@linaro.org>
[catalin.marinas@gmail.com: exported DESTDIR and PYTHON]
Signed-off-by: Catalin Marinas <catalin.marinas@gmail.com>