Fix quilt series import with non-default patch strip levels
This patch fix the failure introduced in the preceding patch adding a test for
quilt series import. It has other safety guards around the same subject, as
quilt can have various "-p" options given to it, but only "-p0" is encoded in
the series file. So try to detect the potentially other failure modes.

Signed-off-by: Vincent Legoll <vincent.legoll@gmail.com>
Signed-off-by: Catalin Marinas <catalin.marinas@gmail.com>