Coverage support
Enable code coverage using the `coverage` package. The top-level stg script
tests for the 'COVERAGE_PROCESS_START' environment variable to enable code
coverage mode.

Coverage is configured to generate separate .coverage files for each stg
process. This strategy allows aggregate coverage to be collected from
running the entire test suite.

A new 'make coverage' directive is added that invokes the test suite with
coverage enabled and collects the per-process .coverage files into a single
top-level .coverage file. An html report is generated.

The test system is modified so that the generated .coverage files are
excluded via the .git/info/exclude file. This required several tests to
append to the exclude file instead of overwriting it.

Signed-off-by: Peter Grayson <jpgrayson@gmail.com>
Signed-off-by: Catalin Marinas <catalin.marinas@gmail.com>