contrib/vim: Fix filetype detection in VIM >=7.4
The command "setfiletype" will not override an existing filetype. This was
never a problem for me previously, but since upgrading from VIM 7.3 to 7.4
the filetype for StGit's files is explicitly set to "text", preventing the
stgit ftdetect plugin overriding it. Use "setlocal filetype=" instead to
ensure that we override any previously detected filetype.

Signed-off-by: Zane Bitter <zbitter@redhat.com>
Signed-off-by: Catalin Marinas <catalin.marinas@gmail.com>