Handle arbitrary annotations when inserting sign-off
The Signed-off-by line should appear at the end of the commit message, and
immediately after any other lines of a similar format that may be used by
various tools - not necessarily just those that are whitelisted at present.
One example is the Change-Id line used by Gerrit to identify code reviews.
Users also frequently add their own informally-defined annotations in the
same format to the same block.

When deciding whether to add a newline between the last paragraph and the
signature, look for whether all lines in the last paragraph are in this
format rather than just searching for known sign-off types in the message.

Signed-off-by: Zane Bitter <zbitter@redhat.com>
Signed-off-by: Catalin Marinas <catalin.marinas@gmail.com>