stgit: Minor fix for pick expose option
I sent the following patch to the stgit-users list back in January, but looks like I forgot
to cc any of you. Would it be possible to get this merged?

Thanks,
Jerry

--8<--

[PATCH] stgit: pick: put imported from commit information on new line

Currently if you pick a commit using the expose option it
appends the imported from commit line immediately after
what it gets for the log without inserting a new line.
This can result in the imported from commit line being
tagged on to the end of a Signed-off-by: line or something
else.

Check to see that message ends with newline, and append
a newline if needed before adding imported from commit
data.

Signed-off-by: Jerry Snitselaar <jsnitsel@redhat.com>
Signed-off-by: Catalin Marinas <catalin.marinas@gmail.com>