Repair custom any() and all() implementations
These functions were only defined in the the stgit.utils module when they
were not present in __builtins__, however transaction.py unconditionally
imported these symbols. This was breaking PyPy since builtins are not
assumed to be in the stgit.transaction module namespace. Apparently in
CPython, builtins are implicitly in modules' namespaces.

Signed-off-by: Peter Grayson <jpgrayson@gmail.com>
Signed-off-by: Catalin Marinas <catalin.marinas@gmail.com>