Fix problems with running hooks on Windows
On Windows we can't simply run the hook script directly, since it's not a
batch file. Instead, explicitly call bash to run the hook.

Also, on Windows NamedTemporaryFile opens the temporary file with the flag
O_TEMPORARY to automatically delete it when it is closed, and that means
that to reopen the same file requires passing O_TEMPORARY again (or, more
specifically, they must open it with the FILE_SHARE_DELETE sharing flag in
the underlying Win32 API. See http://bugs.python.org/issue14243#msg164514
for gory details). To avoid using this flag, pass delete=False and just
clean up the file explicitly when we're done.

Signed-off-by: Zane Bitter <zbitter@redhat.com>
Tested-by: Jamie Madill <jmadill@google.com>
Signed-off-by: Catalin Marinas <catalin.marinas@gmail.com>