Find substring with `in`
Use `sub in s` test instead of `str.find(s, sub) != -1`. Calling the
unbound str.find() method is not idiomatic.

All that said, the main issue with this idiom is that it assumes that the
arguments are going to be of type str which while currently true for Python
2, may not be true when Python 3 is supported since we might be dealing
with bytes, str, or unicode objects.

Signed-off-by: Peter Grayson <jpgrayson@gmail.com>