Normalize __future__ imports, encoding, etc.
We add the magic encoding comments to all .py source files to explicitly
identify that UTF-8 is being used. Some files had the encoding comment, but
now all files have it.

We also normalize the __future__ imports and perform this import in every
.py file.

Since the __future__ imports must precede any other imports, assignments,
or definitions, we also move the __copyright__ strings to always be below
imports.

Signed-off-by: Peter Grayson <jpgrayson@gmail.com>