Replace xrange() with range()
Calls to xrange() are replaced with range() for Python 3 compatibility. For
Python 2, this means that less efficient temporary lists will be
constructed instead of the more efficient iterable returned by xrange.

Since most occurrences of xrange() were bound by the number of patches
(small), it is expected that this will have neglible performance impact
when running with Python 2.

Signed-off-by: Peter Grayson <jpgrayson@gmail.com>