Replace mailbox.UnixMailbox with mailbox.mbox
The mailbox.UnixMailbox class is deprecated in Python2.6 and does not exist
in Python3. The mailbox.mbox class replaces UnixMailbox, but requires a
path to the mbox file, not a file object.

We modify `stg import --mbox` to use a NamedTemporaryFile instead of a
StringIO to support reading an mbox from stdin.

Signed-off-by: Peter Grayson <jpgrayson@gmail.com>