Turn-on unicode_literals everywhere
When running with Python 2, importing unicode_literals from __future__
causes string literals in that module to be `unicode` instead of `str`.

When running with Python 3, this is a no-op since string literals are
already unicode `str`.

Turning-on unicode_literals everywhere makes Python 3 string semantics the
default and thus forces unicode hygiene across the codebase. Without
unicode_literals, it is ambiguous whether any given string variable is
encoded bytes or decoded text.

Here is some discussion on the merits and drawbacks of this
unicode_literals strategy:

  http://python-future.org/unicode_literals.html

Signed-off-by: Peter Grayson <jpgrayson@gmail.com>