Add STGIT_TEST_KEEP_LOCALE to test-lib.sh
The default behavior of the tests remains the same: normalize to using the
POSIX/C locale. However, this new STGIT_TEST_KEEP_LOCALE environment
variable may be used to override this behavior and keep the user's
configured locale settings (LC_ALL).

This is useful for being able to run the test suite with a UTF-8 enabled
locale (such as en_US.utf-8).

Signed-off-by: Peter Grayson <jpgrayson@gmail.com>