Reform sys.argv to be unicode
In Python 2 environments, sys.argv contains bytes items. We normalize
sys.argv to be decoded unicode text as it is in Python 3.

In Python 3, sys.argv is always decoded using the encoding specified by the
locale (e.g. LC_ALL=en_US.utf-8 or LC_ALL=C). In misconfigured
environments, it is possible for the user to specify UTF-8 strings on the
command line when the POSIX (C) locale is enabled. In this situation, when
Python 3 decodes the command line, surrogate escape sequences will be
embedded in the ascii string.

We take a cue from PEP-540 and assume that when the system encoding is
ascii, that what the user really wants is utf-8. Since utf-8 is a superset
of ascii, there is little downside to this assumption. For sys.argv, this
means that we re-decode as utf-8 after encoding with os.fsencode(). We use
os.fsencode() because it correctly encodes the surrogate escape sequences.

Signed-off-by: Peter Grayson <jpgrayson@gmail.com>