config: parse booleans similar to git-config
Currently, stgit has three "boolean" configuration values which we
support, "autoimerge", "keepoptimized" and "refreshsubmodules". However,
we only support "yes" and non-"yes" as valid values. (i.e. either the
value is "yes" and it is deemed true, or it is not "yes").

The standard for git config is to use the "--bool" parameter, which
interprets configuration values as boolean literals:

  "yes", "on", "true" and any positive number are true
  "no", "off", "false", zero, and the empty string are false

  Additionally, a key defined without "= <value>" is taken as true.

Since this is the standard method for git, add a new getbool function
for the config module, which will parse values in this way.

We'll now use this getbool() method for each of these values, ensuring
that these boolean configurations are treated similar to proper git
configuration.

Note that we do not use get() directly in implementing getbool(). This
is because we need to distinguish the case of a defined key with an
undefined value to be True. Since get() converts KeyError exceptions
into None, we would not be able to distinguish the cases of an undefined
key (which should be False) with a key that has an undefined value
(which should be True).

I thought to add new tests for these, but since we don't directly expose
the config.py module, and I don't see an easy way to add tests for
python code explicitly

Signed-off-by: Jacob Keller <jacob.e.keller@intel.com>